fs: Remove compilation warnings
Fix the following compilation warnings when using
arm-dey-linux-gnueabi-gcc (GCC) 5.2.0

fs/namespace.c: In function 'SyS_mount':
fs/namespace.c:2721:6: warning: 'kernel_dev' may be used uninitialized in this function [-Wmaybe-uninitialized]
  ret = do_mount(kernel_dev, kernel_dir->name, kernel_type, flags,
      ^
fs/namespace.c:2700:8: note: 'kernel_dev' was declared here
  char *kernel_dev;
        ^
fs/namespace.c:2721:6: warning: 'kernel_type' may be used uninitialized in this function [-Wmaybe-uninitialized]
  ret = do_mount(kernel_dev, kernel_dir->name, kernel_type, flags,
      ^
fs/namespace.c:2698:8: note: 'kernel_type' was declared here
  char *kernel_type;
        ^
And,

fs/nfs/nfs4proc.c: In function 'nfs4_proc_get_rootfh':
fs/nfs/nfs4proc.c:2969:10: warning: switch condition has boolean value [-Wswitch-bool]
  switch (auth_probe) {
          ^

+++ fs/nfs/nfs4proc.c
@@ -2966,7 +2966,7 @@ int nfs4_proc_get_rootfh(struct nfs_server *server, struct nfs_fh *fhandle,
 {
 	int status;
 
-	switch (auth_probe) {
+	switch ((unsigned short)auth_probe) {
 	case false:
 		status = nfs4_lookup_root(server, fhandle, info);
 		if (status != -NFS4ERR_WRONGSEC)
